+++ server/component/style/formUserInput/FormUserInputModel.php
@@ -340,11 +340,13 @@ class FormUserInputModel extends StyleModel
         $data_config = $this->get_db_field("data_config", '');
         $subject = $this->get_db_field("email_subject", '');
         $body = $this->get_db_field("email_body", '');
+        $email_address = $this->get_db_field("email_address", '');
         if($entry_data){
             // $entry_data = json_decode($entry_data, true);            
             $body = $this->get_entry_value($entry_data, $body);
             $subject = $this->get_entry_value($entry_data, $subject);
             $data_config = $this->get_entry_value($entry_data, $data_config);
+            $email_address = $this->get_entry_value($entry_data, $email_address);
         }
         if ($data_config) {
             $fields = $this->retrieve_data($data_config);
@@ -354,7 +356,8 @@ class FormUserInputModel extends StyleModel
                     $body = str_replace($field_name, $field_value, $body);
                 }
             }
-        }        
+        }
+        $email_address = str_replace('@email_user', $this->db->select_by_uid('users', $_SESSION['id_user'])['email'], $email_address);     
         $mail = array(
             "id_jobTypes" => $this->db->get_lookup_id_by_value(jobTypes, jobTypes_email),
             "id_jobStatus" => $this->db->get_lookup_id_by_value(scheduledJobsStatus, scheduledJobsStatus_queued),
@@ -362,7 +365,7 @@ class FormUserInputModel extends StyleModel
             "from_email" => PROJECT_NAME . '@unibe.ch',
             "from_name" => PROJECT_NAME,
             "reply_to" => PROJECT_NAME . '@unibe.ch',
-            "recipient_emails" => $this->db->select_by_uid('users', $_SESSION['id_user'])['email'],
+            "recipient_emails" => $email_address,
             "subject" => $subject,
             "body" => $body,
             "description" => "FormUserInput Feedback email"