+++ server/db/update_v4.0.0_v4.1.0.sql
@@ -69,7 +69,9 @@ VALUES (get_style_id('tab'), get_field_id('icon'), '', 'Show icon; For web font 
 
 # executed on studybuddy
 
-
+INSERT INTO `fields` (`id`, `name`, `id_type`, `display`) VALUES (NULL, 'email_address', get_field_type_id('text'), '0');
+INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
+VALUES (get_style_id('formUserInput'), get_field_id('email_address'), '@email_user', 'Use `@email_user` to retrive automaticaly the user email. Emails are separated by the MAIL_SEPARATOR. It is `;`');